+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Support for Image activities, namely from Hubzilla
+- Argon2 password hashing
 
 ### Fixed